Move lookup utility class from azure to global project
